Properly(ish) do authentication and get enrollments
A caveat to this: D2L.LP.Tools.Extensibility.WebExtensibility.Cors.CorsHttpModule.HandleCors() doesn't apply CORS headers to the first two requests we're doing here, so you have to modify it to do so. Because of this, you're now running a localhost instance (which isn't considered "valid" in HandleCors()), so you have to disable the .IsValid() check as well. Once you've done this, and you're logged into an LMS, this should all work, and get you your enrollments. Yay!

Oh, also you'll need to set "lms" and "uisEndpoint" to be correct as well.

Of course, none of this is really a problem if you're running the widget on the LMS itself, but, that's not in place yet.