Fetch the largest-required image from Nitro
This ensures that when we fetch the course tile background image from Nitro, it will return a version of the image that has a resolution that matches our max width of the image. If the original image is smaller than 768px wide, this doesn't really do anything, but for larger images we need to specify this, otherwise Nitro returns a 640x480 image.