Shorten transition time
Through many experiments, it was found that the issue seems to be the combination of a subtle change (5% scale) with a long (0.5s) transition. This just makes each keyframe really noticeable, giving the choppiness. Two possible solutions were to increase the change (i.e. higher scale), or shorten the transition; going with the larger scale.