Change demo structure and fix demo bug
index.html is now just a list of WC demos, so that they're more discoverable.

Also had to fix a bug in the touch menu demo. This was occurring because the parent is not a Polymer component - might want to consider updating the menu component so that this is not a requirement.