Fix various bugs
Mostly in animation, that I introduced with these changes. Because there's now a distinction between the enrollment (which is what gets pinned) and the organization (which is what has the name, homepage link, etc.), this complicates things a bit. This does include one hack, but this is going to require a larger refactor to fix; for now, it'll have to suffice ("for now" being as short of a time as possible, preferably).