Fix demos, to a certain degree at least
Updated the d2l-my-courses demo to use the newer data format now that we've switched to the Hypermedia APIs. Removed the d2l-all-courses demo, as there's been rework recently that means it's kinda weird to have it demoed, since it's just a thin layer now.

This is not intended to be final, but just get the demos back to a working state prior to doing any other rework.