Simplify how column recalculation is done
Rather than calling into and out of various elements to manually trigger their `_rescaleCourseTileRegions` method, instead just fire the the `recalculate-columns` event whenever something happens that warrants recalculating how many columns should be displayed. Currently this includes:

- window resize
- receiving enrollments
- opening all courses overlay
- closing all courses overlay

It explicitly does *not* include when a tile is actually moved from one state to the other - the reason being, I'm not sure if this would be a jarring experience. This does mean that, if you're unpinning/unpinning courses, the number of columns won't change (until a refresh, resize, or if you're doing it in the all courses overlay and you close it).