Add delay-load property to course tile, and delay all-courses
When true, delay-load prevents the tile from doing a network call to fetch the organization initially. The tiles can then be loaded by setting delay-load=false.

It is set to true for the all-courses overlay grids, since these have far more enrollments (typically). When the overlay is opened, we load the tiles - thanks to the delay introduced by the fade in of the overlay, even with the 2G network throttling in Chrome the loading is invisible. (On the GPRS setting, it is somewhat noticeable, but they load in order, so it's only if you open the overlay then quickly scroll that you'll notice it.)