Initial use of filtering API
Some hardcoded URLs in here, but the gist of this is that it fetches the list of semesters and creates a dropdown with them. When one (or many) of these options are selected, that is added to the search request.

This also includes the change that the searching now occurs on the /enrollments endpoint, rather than the /organizations endpoint. This came about as a result of an unfortunate gap in what /organizations can support (within reason), but also kinda makes sense, as this isn't intended to be a "search all organizations" thing, more a "search my stuff" thing - so, /enrollments.