Fix bug where semester/department lists didn't load
This was a pretty subtle one - the _selectSemesterList function calls .clear() on the search widgets in ready(), which meant they were getting results back before the handlers were set up. Since .search() on the search widgets only actually does the search if the search term changes (which it didn't), they weren't actually doing the call.

Solution: call _selectSemesterList after the handlers are set up. That took an embarassingly long time to figure out.