Fetch actual department/semester type IDs
What the actual organization type ID of a semester or department is can be set in the LMS (not sure if it ever really changed, but can be). Safest route is to hit the /semester and /department routes in Valence, which give us back the ID we know to be correct, and use that as the organizationTypeId for the department/semester list/search requests.