Move requests for departments/semesters
This call is... fairly expensive, so no need to do it if the user isn't even opening the all courses overlay. Strictly speaking, we could delay it until they actually click the filter dropdown, but a) may be an overoptimization, and b) would need to add a loading spinner, as it takes a few seconds typically.