Re-add network call optimization
Don't bother fetching filter menu contents when user has < 20 enrollments. Also, when the user has more than 20 enrollments, but semesters.length < 2 && departments.length < 2, don't bother populating the menu.

This also switches the filter menu to be default-hidden; otherwise, it briefly appears during the time it takes for the network call to complete, then disappears again. With this method, it does take a moment to show, but it's being loaded async so that's kinda unavoidable.