Clean up d2l-all-courses tests
A lot of functionality was removed... a lot of untested functionality. Yikes. That's on me.

Either way, cleaned up these tests a bunch, removed a bunch of unnecessary stuff, and added a couple for handling the filter-change event that now comes from d2l-filter-menu-content. All the complex behaviour is in there now, so it'll have to be tested there instead.

Note to self: actually do that. You lazy butt.