De-confusing _hasEnrollments + tests
This all started with writing a test to make sure the search disappeared when the user had no enrollments... (it didn't)

This led to reworking how _hasEnrollments and _hasPinnedEnrollments work - now, they're just automatically updated by an observer whenever the pinnedEnrollments or unpinnedEnrollments changes. This has the advantage of moving that calculation into the behaviur itself, which is really where it belongs. This also means it's now working properly in d2l-all-courses.

Also:
- add a couple tests
- don't `this.set('enrollments')` in d2l-my-courses, since it's not used anyway
- rename filter -> filterMenuContent, which is a less- :hankey: name