Re-add lazy load behavior + more tests
Unfortunately it seems that the only way to get this to trigger is for it to have a scroll target of the d2l-dropdown-content-container (in d2l-all-courses), which means d2l-all-courses has to reach into d2l-filter-menu-content. Not great, but tried many other ways and none of them worked. This works. Therefore, this wins.