Create d2l-all-courses on hover/focus
This seems to be a good mix of avoiding creating of d2l-all-courses, while not having it noticeably lag when the user wants to open it. The time taken on a touchpad to lift your finger and tap the link after hovering over it seems to be enough for the creation to run, so there's no really noticeable delay (and when there is, e.g. in FF or IE11, it's only a small amount longer than the already-somewhat-laggy opening of the overlay anyway). In performance testing, this has the same page.display as deferring creation to button click, but has the advantage of less visual lag.