Open overlay first, then set up/load
This further reduces the noticeable delay when opening the overlay the first time. The animation takes 500ms, but using an exactly 500ms delay still results in a bit of lag (just at the end of the animation), so increased this to 600ms to ensure the animation is definitely finished. Also added a loading spinner for this interim time.