Restyle and change search bar's internal behaviour
The search bar as it was previously did not match the behaviour of other search bars in the LMS. In particular, the desired behaviour is such that the state switches between "searched" and "not searched". This is effectively reflected through the state of the button icon - if you have searched for your current term, the button (X) will clear the field, and "redo the search with no query", clearing it back to the "not searched" state. If you change the term (or, in the initial state), the button goes back to the magnifying glass, and activating it (or hitting Enter) does the query and enters the "searched" state.

This also involved some rework of the under-the-hood stuff here, to take fuller advantage of some of Polymer's features as well as address some minor issues that previous existed.

Fixes #247 (kind of a side effect)
Fixes #248 (kind of a side effect)