Step two: don't render d2l-ajax
There's no real need for d2l-ajax to be in the DOM in general, but in this case we were rendering three, two of which will hardly ever be used (pin/unpin and the telemetry request). So, those can definitely be created on-demand. As for the third one (which fetches the organization for the course tile when its enrollment property is changed), it's always needed, but critically - it is only ever needed in our application after at least one network call in the parent (the widget fetches /enrollments), meaning it will never immediately receive the enrollment anyway.

Also changed/simplified the ready callback - it was doing stuff it didn't need to, since we manually call generateRequest only when delayLoad or the enrollment change anyway.

These changes further reduced the polyperf testing, down to <1400ms/250 components.