Step three: dom-if the hover menu
This is a super heavy component, so putting behind a template means it will have largely no cost, until the first time the tile is hovered. While this won't actually make the tile itself much faster or anything, it will further reduce the amount of WC registration/DOM manipulation being done when the tile is first created - a time when many other things are also generally competing for resources. This only-create-on-hover approach allows us to work at the speed of the user instead, which is much slower.