Typography auditing
Turns out a bunch of styles were either being double-applied, not applied, applied in ways that weren't the correct way, or not applied by worked by some random happenstance. These changes remove as much of this uncertainty as possible, by primarily relying on d2l-typography as much as possible, and not modifying those styles unless explicitly required. A lot of small changes here and there, many of which have no visual effect, a few of which have minor visual effects (but these align us to be more consistent with the rest of the LMS, which is using d2l-typography).