Use d2l-search-widget-behavior
The intent here is to simplify d2l-search-widget-custom by relying on the existing behaviour. A few things are overriden from the behaviour (as we have more complex behaviour with this version of it), but it did significantly reduce overlap, and means future updates and changes will generally only have to happen in one place.

Note that a lot of this diff's nastiness is just moving functions around, so that things were easier to find...