Update d2l-simple-overlay
Previously, we were "accidentally" getting access to the web animation API as the simple overlay used neon animation, which used the web animation polyfill. The new version removes the dependency on neon-animation, so we no longer get it for free. Should have been explicitly including it in the first place anyway, really.