Refactor course tile (un)pin animations
Primary motivation was to remove web-animations-js as a dependency, the
simplest act then would've been to replace to the calls to #animate()
with a similar series of modifications and event listeners; however, in
trying to understand the whole procedure, I also poked at it. The
overall set of changes are arguably "better", but probably mostly just
different :).

Creating the animations unfortunately still revolve around "quick do
things before the next paint", which means we've cornered ourselves into
forced re-calc/layout territory. Believe with further refactoring of
what's causing the full set of animations, we should be able to fix this
up.

Also, the current animation code appears broken in Chrome 57 so - fixed!

* remove web-animations-js usage
* move (un)pin animation behaviour, into its own behaviour
* remove "queue" logic. while more correct, seems unlikely to be needed
* base tile movement on maths and the size of the given tile. overall
  less accurate than client rectangles, but should be "faster"?
* support multiple tile insertions/removals, if that's ever a thing
* various tweaks in relation

Ref: US82506