Fix #272
The crux of the console errors we were seeing was that the d2l-search-dropdown-content was not quite exactly the same as d2l-dropdown-content, while still using the DropdownContentBehavior. There really wasn't much benefit to d2l-search-dropdown-content anyway, so it's been removed, and the console errors are no more.

Also, switch the Polymer-2.0-friendly "slot" while we're here.

Also also, did some changes to visual consistency so that the live search dropdown matches the Sort By dropdown formatting in terms of text and style (confirmed with Bob that this is the correct approach).

Also also also, added a call to _updateIcon() as selecting an item in the live search list would result in a situation where the results couldn't be cleared.