Force filter menu to resize after opening
If the list of semesters or departments is long, for some reason it will just extend past the bottom of the viewport rather than resizing correctly. To force it to size correctly, we can just fake-trigger a resize event, which causes magic to happen deep within d2l-menu (or maybe d2l-dropdown (or maybe both)) and it resizes to the right size. This isn't a hack, it's more just a... suboptimal solution, but I've spent hours trying to figure out another way, and not been able to really identify a "cleaner" way to do it.